+++ src/JsonApi/Exception/FilteringCriteriaUnsupported.php
@@ -1,0 +1,28 @@
+<?php
+namespace WoohooLabs\Yin\JsonApi\Exception;
+
+class FilteringCriteriaUnsupported extends \Exception
+{
+    /**
+     * @var string
+     */
+    private $criteria;
+
+    /**
+     * @param string $criteria
+     */
+    public function __construct($criteria)
+    {
+        parent::__construct("Filtering criteria '" . $criteria . "'' is unsupported!");
+
+        $this->criteria = $criteria;
+    }
+
+    /**
+     * @return string
+     */
+    public function getCriteria()
+    {
+        return $this->criteria;
+    }
+}