+++ src/JsonApi/Transformer/CollectionFilterTrait.php
@@ -1,6 +1,7 @@
 <?php
 namespace WoohooLabs\Yin\JsonApi\Transformer;
 
+use WoohooLabs\Yin\JsonApi\Exception\FilteringCriteriaUnsupported;
 use WoohooLabs\Yin\JsonApi\Exception\SortingCriteriaUnsupported;
 use WoohooLabs\Yin\JsonApi\Schema\Included;
 
@@ -10,6 +11,7 @@ trait CollectionFilterTrait
      * @param array $data
      * @param \WoohooLabs\Yin\JsonApi\Schema\Included $included
      * @param array $sortingFields
+     * @return array
      */
     protected function sortByFields(&$data, Included $included, array $sortingFields)
     {
@@ -20,7 +22,7 @@ trait CollectionFilterTrait
                     $b = $this->getResourceField($b, $field, $included);
                 }
 
-                $result = $this->compare($sorting["field"], $a, $b);
+                $result = $this->performSortingComparison($sorting["field"], $a, $b);
                 if ($result !== 0) {
                     return $result * $sorting["direction"];
                 }
@@ -33,12 +35,77 @@ trait CollectionFilterTrait
     }
 
     /**
+     * @param array $data
+     * @param \WoohooLabs\Yin\JsonApi\Schema\Included $included
+     * @param array $filteringFields
+     * @return array
+     */
+    protected function filterByFields(&$data, Included $included, array $filteringFields)
+    {
+        $filter = function ($item) use ($filteringFields, $included) {
+            foreach ($filteringFields as $filtering) {
+                foreach ($filtering["field"] as $field) {
+                    $item = $this->getResourceField($item, $field, $included);
+                }
+
+                $result = $this->performFilteringComparison(
+                    $filtering["originalField"],
+                    $item,
+                    $filtering["operator"],
+                    $filtering["value"]
+                );
+                if ($result === false) {
+                    return false;
+                }
+            }
+
+            return true;
+        };
+
+        // Filtering primary data and collecting relationship identifiers which can be included
+        $remainingIncludableRelationships = [];
+        foreach ($data as $key => $resource) {
+            if ($filter($resource) === false) {
+                unset($data[$key]);
+            } else {
+                $this->addRelationshipIdentifiers($remainingIncludableRelationships, $resource);
+            }
+        }
+
+        // Filtering included data
+        $included->filterResources(function ($type, $id) use ($remainingIncludableRelationships) {
+            return isset($remainingIncludableRelationships[$type][$id]) ? true : false;
+        });
+    }
+
+    /**
+     * @param array $relationships
+     * @param array $resource
+     */
+    protected function addRelationshipIdentifiers(&$relationships, $resource)
+    {
+        if (isset($resource["relationships"]) === false) {
+            return;
+        }
+
+        foreach ($resource["relationships"] as $relationship) {
+            if (isset($relationship["data"]) && $this->isAssociativeArray($relationship["data"]) === false) {
+                foreach ($relationship["data"] as $item) {
+                    if (isset($item["type"]) && isset($item["id"])) {
+                        $relationships[$item["type"]][$item["id"]] = true;
+                    }
+                }
+            }
+        }
+    }
+
+    /**
      * @param mixed $resource
      * @param string $field
      * @param \WoohooLabs\Yin\JsonApi\Schema\Included $included
      * @return string|null
      */
-    private function getResourceField($resource, $field, Included $included)
+    protected function getResourceField($resource, $field, Included $included)
     {
         if ($resource === null) {
             return null;
@@ -66,11 +133,41 @@ trait CollectionFilterTrait
     /**
      * @param string $fieldName
      * @param mixed $x
+     * @param string $operator
+     * @param mixed $y
+     * @return bool
+     * @throws \WoohooLabs\Yin\JsonApi\Exception\FilteringCriteriaUnsupported
+     */
+    protected function performFilteringComparison($fieldName, $x, $operator, $y)
+    {
+        if (($x !== null && is_scalar($x) === false) || ($y !== null && is_scalar($y) === false)) {
+            throw new FilteringCriteriaUnsupported($fieldName);
+        }
+
+        switch ($operator) {
+            case "=":
+                return $this->equals($x, $y);
+            case "<":
+                return $x < $y;
+            case "<=":
+                return $x <= $y;
+            case ">":
+                return $x > $y;
+            case ">=":
+                return $x >= $y;
+        }
+
+        return false;
+    }
+
+    /**
+     * @param string $fieldName
+     * @param mixed $x
      * @param mixed $y
      * @return int
      * @throws \WoohooLabs\Yin\JsonApi\Exception\SortingCriteriaUnsupported
      */
-    private function compare($fieldName, $x, $y)
+    protected function performSortingComparison($fieldName, $x, $y)
     {
         if (($x !== null && is_scalar($x) === false) || ($y !== null && is_scalar($y) === false)) {
             throw new SortingCriteriaUnsupported($fieldName);
@@ -89,6 +186,26 @@ trait CollectionFilterTrait
         }
 
         return $this->compareString($x, $y);
+    }
+
+    /**
+     * @param mixed $x
+     * @param mixed $y
+     * @return bool
+     */
+    protected function equals($x, $y)
+    {
+        if (is_bool($x)) {
+            $y = boolval($y);
+        } elseif(is_long($x)) {
+            $y = intval($y);
+        } elseif (is_double($x)) {
+            $y = doubleval($y);
+        } elseif (is_string($x)) {
+            $y = print_r($y, true);
+        }
+
+        return $x === $y;
     }
 
     /**
@@ -119,5 +236,14 @@ trait CollectionFilterTrait
     private function compareString($x, $y)
     {
         return strnatcasecmp($x, $y);
+    }
+
+    /**
+     * @param array $array
+     * @return bool
+     */
+    private function isAssociativeArray(array $array)
+    {
+        return (bool)count(array_filter(array_keys($array), 'is_string'));
     }
 }